Track dictionary

+++ src/datahawk/import_dialog.py
@@ -8,9 +8,9 @@ from PySide6.QtWidgets import (
 )
 
 from datahawk.source.mychron.mychron import check_device, list_sessions, download_session, Session
-from datahawk.storage import save_session, get_imported_filenames
+from datahawk.storage import save_session, get_imported_filenames, load_track, save_track
 from datahawk.source.mychron.xrz_parser import parse_xrz
-from datahawk.session_processing import process_session
+from datahawk.track_selector import TrackSelector
 
 
 class _ListWorker(QThread):
@@ -32,30 +32,37 @@ class _DownloadWorker(QThread):
     finished = Signal(int)
     error = Signal(str)
 
-    def __init__(self, sessions: list[Session], driver: str):
+    def __init__(self, sessions: list[Session], driver: str, track_name: str):
         super().__init__()
         self._sessions = sessions
         self._driver = driver
+        self._track_name = track_name
 
     def run(self):
         try:
             count = 0
+            track_created = False
             for i, s in enumerate(self._sessions):
                 self.progress.emit(i + 1, len(self._sessions))
                 data = download_session(s.name, expected_size=s.size)
                 if data and len(data) > 200:
-                    # Detect best lap time from the session data
-                    try:
-                        from pathlib import Path
-                        import tempfile
-                        tmp = Path(tempfile.mktemp(suffix='.xrz'))
-                        tmp.write_bytes(data)
-                        parsed = parse_xrz(tmp)
-                        session = process_session(parsed)
-                        blt = session.best_lap_time
-                        tmp.unlink()
-                    except Exception:
-                        blt = None
+                    # Create track from first session if it doesn't exist yet
+                    if not track_created and not load_track(self._track_name):
+                        try:
+                            import tempfile
+                            from pathlib import Path
+                            from datahawk.session_processing import detect_sf_line, detect_master_lap
+                            from datahawk.types import Track
+                            tmp = Path(tempfile.mktemp(suffix='.xrz'))
+                            tmp.write_bytes(data)
+                            parsed = parse_xrz(tmp)
+                            sf_line = detect_sf_line(parsed)
+                            master_lap = detect_master_lap(parsed, sf_line)
+                            save_track(Track(name=self._track_name, sf_line=sf_line, master_lap=master_lap))
+                            tmp.unlink()
+                        except Exception:
+                            pass
+                    track_created = True
 
                     save_session(
                         driver=self._driver,
@@ -64,8 +71,8 @@ class _DownloadWorker(QThread):
                         date=s.date,
                         time=s.time,
                         laps=s.laps,
-                        track=s.track,
-                        best_lap_time=blt,
+                        track=self._track_name,
+                        best_lap_time=None,
                     )
                     count += 1
             self.finished.emit(count)
@@ -90,6 +97,11 @@ class ImportDialog(QDialog):
         self._driver_input.setPlaceholderText("Enter driver name")
         driver_row.addWidget(self._driver_input)
         layout.addLayout(driver_row)
+
+        # Track selection
+        self._track_selector = TrackSelector()
+        self._track_selector.changed.connect(self._update_import_btn)
+        layout.addWidget(self._track_selector)
 
         # Status bar
         status_row = QHBoxLayout()
@@ -136,7 +148,8 @@ class ImportDialog(QDialog):
     def _update_import_btn(self):
         has_selection = len(self._table.selectedItems()) > 0
         has_driver = bool(self._driver_input.text().strip())
-        self._import_btn.setEnabled(has_selection and has_driver)
+        has_track = bool(self._track_selector.track_name)
+        self._import_btn.setEnabled(has_selection and has_driver and has_track)
 
     def _load_sessions(self):
         self._status.setText("Connecting to device...")
@@ -191,7 +204,7 @@ class ImportDialog(QDialog):
         self._progress.show()
         self._status.setText("Downloading...")
 
-        self._dl_worker = _DownloadWorker(selected, driver)
+        self._dl_worker = _DownloadWorker(selected, driver, self._track_selector.track_name)
         self._dl_worker.progress.connect(self._on_dl_progress)
         self._dl_worker.finished.connect(self._on_dl_done)
         self._dl_worker.error.connect(self._on_dl_error)
